backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
